cloud_functions/twitter_ipl_teams_tweets/requirements.txt
google-cloud
google-cloud-storage
pandas